Filter out signatures with no country when resetting journal
This restores parity between the `CountryPetitionJournal` methods `.record_new_signature_for` and `.reset!`.  Our schema allows for null countries on signatures even if our models (currently) don't so we have to jump through hoops a bit in our tests for this.  Our SQL also has to try to do something similar to Rails' `String#blank?` to filter things out.  It shouldn't be possible, but it's better to be safe.